Pace burger icon and chwl logo on right place, move navbar component to navbar folder
