+++ src/components/navbar/Navbar.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import DesktopNav from "./navbar/ApplicationHeader";
-import MobileNav from "./navbar/MobileApplicationHeader";
+import DesktopNav from "./ApplicationHeader";
+import MobileNav from "./MobileApplicationHeader";
 
 const Navbar = () => {
   const mql = window.matchMedia("(max-width: 1200px)");
